+++ projects/igniteui-angular/src/lib/icon/index.ts
@@ -2,23 +2,14 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
 import { IgxIconComponent } from './icon.component';
-import { IgxIconService } from './icon.service';
 import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
     declarations: [IgxIconComponent],
     exports: [IgxIconComponent],
-    imports: [CommonModule, HttpClientModule],
-    providers: [IgxIconService]
+    imports: [CommonModule, HttpClientModule]
 })
-export class IgxIconModule {
-    public static forRoot() {
-        return {
-            ngModule: IgxIconModule,
-            providers: [IgxIconService]
-        };
-    }
-}
+export class IgxIconModule {}
 
 export * from './icon.component';
 export * from './icon.service';