+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes for each version of this project will be documented in this 
 ## 6.2.2
 - `igx-checkbox`:
     - Added a new input property - `disableTransitions`. It allows disabling all CSS transitions on the `igx-checkbox` component for performance optimization.
+- `IgxIconModule`:
+    - **Breaking change** `igxIconService` is now provided in root (providedIn: 'root') and `IgxIconModule.forRoot()` method is deprecated.
+    - **Breaking change** `glyphName` property of the `igxIconComponent` is deprecated.
 
 ## 6.2.1
 ### Features